fix: rename menu and home components
