+++ src/constants/menu-sections.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import i18n from '@dhis2/d2-i18n'
 import UserSyncTestContainer from '../components/sections/user-sync-test/user-sync-test-container'
-import Home from '../pages/Home/home'
+import Home from '../pages/Home/Home'
 import GeneralSettings from '../pages/General/GeneralSettings'
 import GlobalSettings from '../pages/Synchronization/Global/GlobalSettings'
 import ProgramSyncSettings from '../pages/Synchronization/Programs/ProgramSyncSettings'
